Restore the ability to bootstrap Ivy with a custom configuration file.
This is required when publishing with an Ivy plugin. For example, if Ivy is
configured to publish via SVN, the fm.last.ivy.plugins.ivysvnresolver plugin
will need to be loaded into Ivy.  If this is added to the main
ivysettings.xml, Ivy bootstrap will fail.  The solution is to let Ivy
bootstrap itself as per-usual, and at publish time, substitute in a custom
ivysettings.xml file that reverences (the already bootstrapped) Ivy+plugins.

Reviewed at https://rbcommons.com/s/twitter/r/1709/