Remove old publishing settings.
Zinc's distribution (zip/tgz files) used to be published to Typesafe's
Artifactory repository. However, f847e450f6a8eb66d27a1f20fc942ae78278750e
introduced publishing to S3 and the old publishing mechanism was not used
anymore but its config was left in place. To verify that old mechanism
was not used anymore check contents of this repository:

http://typesafe.artifactoryonline.com/typesafe/zinc/com/typesafe/zinc/dist/

The last release published there was done in Dec 2013.

The old config was not used because `publishTo` was scoped to `dist`
subproject. However, the root project had `publishTo` set to Maven Central
and that one was used.

In #62 refactoring, the `dist` got merged into root project and the old
config got reactived. Let's remove this config for good.