Add an 'advanced' parameter to registering options
Advanced options to not normally display with help messages.  They are usually
intended to be set for the entire repo in pants.ini and shouldn't be specified
on the command line because they are difficult to configure or may affect
the hermeticity of the build.

Tagged --target and --source as advanced options, since changing the --source argument
changes what the javac compiler will accept and affects the compiler output.

Testing Done:
Added unit test for PantsHelpFormatter change

./pants --help-all
 output of --target and --source for compile.java is suppressed

./pants --help-all --help-advanced
./pants compile foo --help --help-advanced
  output of --target and --source display as follows:

```
  (ADVANCED)
  --compile-java-source=_COMPILE.JAVA_SOURCE__
  --source _COMPILE.JAVA_SOURCE__
                          Provide source compatibility with this release.
                          (default: 6)
  (ADVANCED)
  --compile-java-target=_COMPILE.JAVA_TARGET__
  --target _COMPILE.JAVA_TARGET__
                          Generate class files for this JVM version. (default:
                          6)
```

Bugs closed: 1062

Reviewed at https://rbcommons.com/s/twitter/r/1739/