Add progress information to partition reporting.
When compiling many partitions, it is nice to be able to determine the
progress without resorting to manually scrolling through the log and
counting the number of completed partitions.

Testing Done:
Compiled a set of pants targets and saw that the progress messages were printed out.

Reviewed at https://rbcommons.com/s/twitter/r/1749/