Fix TestAndroidBase task_type override miss.
The override attempt was against a classmethod in TaskTest but itself
was an instance method.  This was harmless, but also un-needed.  The
TaskTest base class raises and this is a better default, forcing the
TestAndroidBase leaves to implement.

Testing Done:
CI went green here:
  https://travis-ci.org/pantsbuild/pants/builds/50286508

Bugs closed: 1080

Reviewed at https://rbcommons.com/s/twitter/r/1751/