Pull Android xml parsing into AndroidManifestParser class.
This moves all xml parsing out of the Android target. Any method
that retrieves a value from the AndroidManifest.xml is now a cached
property in the android_target.

The methods that parse the xml are now robust against missing values
and ill-formed xml. This class mimics the android_distribution class
in that as long as it is able to parse and find values, those values
are sucessfully returned. It will be up to the consumer to decide
if the value is sufficient for its purposes.

This has a complete test class for the AndroidManifestParser.
An upcoming CR will expand the testing of Android targets and tasks
that consume values from this parser.