Enable python-eval for pants source code.
This installs the PythonEval task in the compile phase for pants and
engages the python "compile" checks for pants source code in ci.  As a
result there are many missing BUILD dep fixes as well flagged by
python-eval in ci.

The python-eval checks add ~2 minutes to each ci shard, so ci is
restructured to be more fine grained in order to keep ci times ~6
minutes per shard.