Fixup IvyResolve integration test.
The new pex honoring of sys.excepthook comes with a twist.  For
un-determined reasons, the traceback handed to our excepthook by
pex does not carry code context lines info and so the backtrace omits
these lines.  Although the backtrace is still useful, it will not
contain the `raise` line where the exception originates from and the
presence of this line was dependend upon by the IvyResolve integration
test.  Update the test to look for the cycle error message instead for
now.