Add XmlParser class and rework AndroidManifestParser.
All the parsing is now done at one time, with XmlParser superclass.
AndroidManifestParser takes the parsed xml and then just pulls the
elements and attributes it needs. XmlParser narrows the
scope of ill-formed xml. The minidom package is a frontend for
several parsers so this catches any ill-formed xml. The
AndroidManifestParser is only responsible for making sure the
necessary elements/attributes are in the parsed output.