Extract scrooge tasks to contrib/.
This establishes a contrib tree for pants plugin contributions as
discussed here:
https://groups.google.com/d/topic/pants-devel/9EREtXy5lR4/discussion

In order to support contribs, a few changes are made to CI:
+ break out a contrib test section + flag and dedicate a travis shard
+ break out the empty __init__.py check to a script and allow for
  namespace declarations.

Also, modify the GoalRunner to support namespace packages when adding to
the `sys.path`.

The release script is ammended to hook into
`contrib/release_packages.sh` but more work needs to be done to the
`setup-py` goal to avoid re-packaging in-repo dependency source (the
pants code itself).