Expand AndroidManifestParser and move most of the cached properties out of android_target.
Create wrapper functions that reduce the complexity of the error checking.
Add a TestXmlBase class that both TestXmlParser and TestAndroidXmlParser inherit from.
All the cached properties that were in AndroidTarget are now mostly
handled in AndroidManifestParser. When a Task needs any property
that is set diretly from the AndroidManifest.xml, it will call
target.manifest.$attribute.