Create a directory with no write perms for TestAndroidConfigUtil.
The original test assumed we would have no write permissions outside
of home, which turned out not to be a safe assumption.

I added the IOError to the exception catch in AndroidConfigUtil.
The OSError is raised from the safe_open, if there is no
permission to make dirs. The IOError is for cases where the
directories exist but pants does not have permission to write there.