Android backend now exclusively uses the new option system.
No arbitrary config access, no CONFIG_SECTION etc.

I do not define a default for the keystore_config_location in the
task because we actually produce a default file. If the keystore_location
is changed, I don't want to fallback to a default silently or
else the user may not know that their config wasn't used.