Setup logging before plugins are loaded.
This lifts logging options to the bootstrap stage in order to be able
to configure logging before plugins are loaded.  The logging setup is
lifted out from GoalRunner to a top-level function in the logging
package and tests are added for logging configuration.

Additionally, bootstrap option parsing is fixed to handle short flags
since logging uses a few of these.