Add pants_bootstradir and pants_configdir to options bootstrapper.
This makes these values available to all classes plumbed with
the options system. The only places I see the boostrapdir used
are in binary_utils and in the ivy boostrapper. So as they config
access is removed from those classes, then they can convert
to getting the pants_bootstrapdir form the options.

I removed the expanduser call from the ivy class as it is now
redundant. The expanduser happens in the get_cachedir() call
that sets pants_bootstrapdir.