Fixup Config to find DEFAULT values for missing sections.
This supports the spirit of the `Config.get` and `Options` docs.
Previously the DEFAULT section was only consulted when the section
being queried existed but had no option defined.  Now the DEFAULT
section is also consulted when the section being queried is not
defined at all.  This should support more less brittle configs that
can have global defaults defined for all current and future scopes.