Add a migration config and notes for ng_daemons.
This necessitated a feature addition and a bugfix in the migrate_config
tool.

To support migrating from a global option to a local one, a
`None` migration destination is now supported to indicate there is no
migration path.  If there is a note for these cases it is displayed;
otherwise the fact there is no migration path is noted.

To support detecting legacy boolean config value properly, fixup the
`has_non_default_option` helper in `check_config_file`.