Improve AndroidDistribution class and coverage.
Inspired somewhat by CR 1850, the caching
of AndroidDistribution has been improved to
where the calls to the os module are no longer
repeated. I also added a second, more specific custom
exception.

On the testing side, this increases test coverage.
This changes TestAndroidBase to TestAndroidMixin.
The TestAndroidBase was a super class
to non-task test classes, classifying it as a
mixin is a better fit.