Remove android_dist from all Android ctors.
As with all good ideas, everything is simpler once this
was added. All Android tasks are directly using the
android_sdk property from AndroidTask to get tools.

This removes the idea of an empty SDK path. When a tool
is requested, the path is instantiated and the tool is
looked for. If it cannot be found, an exception is raised.
This is an intuitive and safe way to handle that and
an improvement over previous handling.