Add Scm.add, change publish to add pushdb explicitly, move scm publish around
If a provided artifact hasn't been pushed before, the pushdb file for it may not be tracked by git yet.
This changes publishing so that the pushdb will be added before committing, ensuring that the commit will contain the pushdb changes.
It also moves the scm publish steps into ScmPublish so that JarPublish needs to know less about publishing to scm.