To add android.jar to the classpath, create a hard link under task's workdir.
The classpath refactor requires that all classpath entries be under the
buildroot. The aapt_gen task was creating a synthetic target that added
android.jar from the local Android SDK. That failed with the new rule.

This adds the register_tool_link method, which creates a hard link
under the task workdir. I could not use a symlink (as suggested by Stu)
because it crashes during the creation of the symlink farm.

I am throwing this up for feedback. Maybe you would prefer I fetch
the jar through ivy or something.