Fix `Git.detect_worktree` to fail gracefully.
Previously if there was no git executable on the path this method
would fail instead of returning `None`.  The upshot was that pants could
not be used outside a git repository.

This also makes `Git.detect_worktree` symmetric with `Git.__init__`,
allowing the git binary name/path to be passed in directly.