walk synthetic targets dependencies when constructing context.target()
A recent change made it so that synthetic targets dependencies aren't traversed which caused their 3rdparty deps to be not added to the classpath in certain cases. This caused compilation errors. This patch retains the derivation checks from the previous one, but expands the synthetic targets after determining which ones are needed.