Remove a spurious import in test_antlr_builder.
The import was supposed to come after the generated code imports, to 'prove'
that namespace packaging works. However isort moved it to the wrong place
for that, and in any case the builder now verifies that namespace packages
are indeed functional as such, so this 'test' is not needed.

This test is deeply flawed anyway. All it tests is that it can import
the generated code from its dependencies, so, for example, it can't
detect a break in the code generation until you clean and regenerate.
Just rerunning the test isn't enough.

Hopefully we'll soon rewrite antlr/thrift python generation uniformly with
the JVM side, and we can get rid of these tests entirely.

Testing Done:
Ran that test.

Also verified that the test didn't work before anyway.

Reviewed at https://rbcommons.com/s/twitter/r/1951/