Prepare the 0.0.31 release.
This is a quick follow-on to 0.0.30 that picks up a fix for ScalaCompile
breaking repos with scalac_plugin targets.

I went ahead and removed the deprecated Provides task since it has been
broken for a long time, but pushed forward all other deprecations a tick
since there was such a short period between releases.