always use relpath on --test file args, refactor execute in junit_run, add more integration tests
Specifying the test to run prefixed with ./ fails because the ./ causes the src path to not match the paths in the classes_by_source product. This changes that so it will always create the relpath when trying to derive the classes.

It also refactors execute to reduce nesting and clarify some of the flag / config interactions.