Adding a --test-junit-coverage-jvm-options flag that allows specifying JVM options for the processes that run the test coverage tools.
Cobertura's "instrument" target runs out of PermGen space when it needs to process a target with a larger number of files, and currently there's no way to increase the default (512M).

Not sure who the right reviewers are, so adding some people that I've talked to in the hipchat room.

Testing Done:
Tested the change on a Twitter target that failed with the default options, and passed with --test-junit-coverage-jvm-options=-Xmx4g --test-junit-coverage-jvm-options=-XX:MaxPermSize=1g

build-support/bin/ci.sh passed.

Reviewed at https://rbcommons.com/s/twitter/r/1968/