Fixup minified_dependencies.
This change fixes the core of the SetupPy minified_dependencies and adds
a hack dependency to each of pants 3 auxiliary sdist targets to produce
correct sdists with no repeated code and an exact set of setup_requires.

The SetupPy test is moved to its proper parallel test tree home and add
a test for the pantsbuild.pants contrib sdist case is added that fails
before the fix.

A follow-on change will remove the need for the 3 hack deps by imposing
a well defined ownership model for python targets that don't provide.