Rid pantsbuild.pants of an un-needed antlr dep.
The dep is only needed by 1 test that attempts to use the generated code
to parse via the antlr runtime.  This change lifts the dep from the
requirements.txt file used to bootstrap pants into a stand-alone
python_requirement_library that the test can depend on without forcing
each pants bootstrap to fetch this dep.

Since 3rdparty/python/BUILD was used for an include in the docs, doc gen
broke leading to fixes for CI doc gen errors.