Implement an exported ownership model.
This implements a full consistent publishing model for setup-py whereby
a target that is not exported must have a single exported target owner.
It also adds support for finding such owners even when the owner is not
in the current target closure to effect consistent, complete recursive
setup-py publishing that does not require any dependency hacks.

As a result several BUILD dep hacks for exported plugins are removed
and the release script is updated to take advantage as well.