Place the workdir below the pants_workdir
Locate the TaskTestBase workdir below the pants_workdir, which is automatically cleaned up.

This makes it easier to test Tasks that create synthetic targets, as it is common/best-practice to create synthetic targets within your workdir. Previously this would have failed indicating that the SourceRoot for the synthetic target was outside the BuildRoot.

Testing Done:
https://github.com/pantsbuild/pants/pull/1346

Reviewed at https://rbcommons.com/s/twitter/r/2007/