Kill pantsbuild dependence on maven.twttr.com.
The maven artifacts pants needs that are not hosted on maven central are
now hosted on bintray at https://dl.bintray.com/pantsbuild/maven.

The https://github.com/pantsbuild/maven-repo repo contains the script to
sync new artifacts to bintray as well as a script that makes it easy to
grab a graph of targets from maven.twttr.com.