Rename Address.path_spec_path to fully_qualifed_spec_path
The impetus for this was the stale comment on target.id
and target.identifier. The docstring read that the
target.id is safe for use as a unix path name, which is
not currently true.

The id is set using address.path_spec_path which processes
the spec_path kwarg in the Address ctor. It joins the
path and the name and converts the os.seps to periods. I
renamed the path_spec_path to fully_qualified_spec_path
since it looks like a package name to me.

I also removed a couple unused imports. I could be convinced
to put that in a separate CR if you'd like.