docs: update comment on target.id
The docs and the methods name conspired to create
the illusion that this returns a path, when it actually
returns the path from the build root and the target name,
joined with periods. This updates the docs for methods
involved.

I also cleaned up the target BUILD definition and removed
a couple unused imports.