Prepend a hash of the spec to path_safe_spec
This hashs the spec and prepends the a 12 digit suffix of that
hash to the beginning of the path_safe_spec. Eric had come
up with an example where simply replacing the seps with periods
returned colliding file names (see the included test case).

Some of the original suggestions to fix this were to ban periods
in target names or to transform the ':' in the spec to a already
banned char. That caused a failure when codegen classes would
use the path_safe_spec as a synthetic target name. There were
ways around this, but in the end I would rather not ban periods
anyway, the pants examples has instances of targets with periods,
which means they probably exist in the wild too.

The solution here is to take a hash of the original spec
and prepend that to the transformed spec. In order to collide, the
specs would have to not only conform to Eric's example but also
they would have to produce the same hash. This is memoized for
good measure. I believe this should be statistically safe.
I used a 12 bit suffix for aesthetic purposes, I could be
induced to just use the entire hash.