Part 2 of 2. Add android_library support.
The is the 2nd half of the android_library support.

Thus revamps aapt_gen and moves it off of the codegen backend.
AndroidBinary targets can depend on any number of AndroidLibrary
targets. Those AndroidLibrary targets are defined as a collection
of a jar files and resources files. These libraries are non-exportable.

The AndroidLibrary will not usually have a target sdk defined in the
manifest. On those times, the AndroidLibraries R.java is compiled
against the target SDK of the dependee binary.

The new aapt_gen defines the gentarget as an AndroidBinary. It
collects all AndroidLibrary deps. Then every AndroidResource dep
in the closure of each AndroidLibrary is processed in order
to create the libraries R.java.

If an android_library is depended upon by 2 different binary targets
of different SDK versions, then Pants needs to create 2 different
R.java for that library, one for each SDK.

As of right now there is no invalidation framework for AaptGen.
That will have to be a TODO for either after simple_gen or after
the needed BUILD files rework, which ever comes first.

There is also two new resolvers in ivy. One is for checked-in jars
under 3rd party. The other is to interact with the .m2 repo in the
Android SDK. That second repo clashes with the new rule that
classpath members must be under the buildroot. I am including it here
in the hopes someone has an easy fix for that.