convert all % formatted strings under src/ to str.format format
For the most part this is a straight forward conversion. The only tricky part is where %d was used on arguments that may be floats. Currently, python's float doesn't support the d format, so I had to change those locations so that the argument was coerced to an int before rendering.