ensure source_roots always loaded; validate source_roots dir is subdir
This changes address mapping so that when an address is mapped, and the BUILD file for it is parsed, it ensures all its parent BUILD files have also been parsed. I put it in the address mapper because it already caches the results of parses, which'll reduce re-excution of build files. Also source_root primarily affects addressable entities.

What I'd really like to do is to remove all side effecting things from build files, and have a cache of all the interesting things from them. Instead of only retaining addressables from build parses, we'd collect source_roots etc. Then that service would be passed into others like the address mapper which could perform tasks and build derivative data structures based on the contents.

- fix maven_layout tests so they exercise the maven_layout function
- raise errors when source roots are outside the BUILD dir they are defined in
- raise errors if a context aware factory fails to create a callable
- add method to BuildFile to get direct parent without finding all ancestors