Add optional support for auto-shading jvm tools.
When registering a jvm tool, the main classname of the tool can be
specified to trigger auto-shading of the tool classpath.

This functionality is dogfooded by JUnitRun for its junit-runner tool.

This changes adds a test for bootstrap_jvm_tools.py that exercises the
normal (legacy) and shading paths as well as updating the
jvm_tool_task_test_base.py infra to work with this new shading option.
The latter enables the pre-existing JUnitRunTest to operate, now using
an auto-shaded jar.