Introduce a jvm binary shader.
This tool initially handles just binary jar auto-shading to support
shading of bootstrapped jvm tools.  It should be easy to extend to add
support for fully customizable non-binary jar shading though.  The tool
uses jarjar to perform the shading but the interface steers clear of
jarjarisms so this should be a stable base to move forward with even if
we end up needing to swap out the shading backend.

Testing Done:
The jarjar jar (unused in this RB, used in next) was synced to bintray from
this commit: https://github.com/pantsbuild/maven-repo/commit/782e0cb4
Tested the sync with a successful resolve:
```console
$ PANTS_DEV=1 ./pants resolve.ivy --open //:jarjar
```

And coverage is good for the new code:
```
$ PANTS_PY_COVERAGE=modules:pants.java.jar.shader PANTS_DEV=1 ./pants test tests/python/pants_test/java/jar/:shader
...
21:53:58 00:00   [test]
21:53:58 00:00     [run_prep_command]
21:53:58 00:00     [test]
21:53:58 00:00     [pytest]
21:53:58 00:00       [run]
                     ============== test session starts ===============
                     platform linux2 -- Python 2.7.8 -- py-1.4.26 -- pytest-2.7.0
                     rootdir: /tmp, inifile:
                     plugins: timeout
                     collected 4 items

                     ../../../../../tmp ....
                      coverage: platform linux2, python 2.7.8-final-0 -
                     Name                               Stmts   Miss Branch BrMiss  Cover
                     --------------------------------------------------------------------
                     src/python/pants/java/jar/shader      86      8     30      8    86%

                     ============ 4 passed in 5.17 seconds ============
                     Name                               Stmts   Miss Branch BrMiss  Cover
                     --------------------------------------------------------------------
                     src/python/pants/java/jar/shader      86      8     30      8    86%

21:54:04 00:06     [junit]
21:54:04 00:06     [specs]
               SUCCESS
```

CI went green here:
  https://travis-ci.org/pantsbuild/pants/builds/57977570

Bugs closed: 663, 1362, 1388

Reviewed at https://rbcommons.com/s/twitter/r/2050/