Fix BootstrapJvmTools to plumb jvm options to Shader.
This was the root cause of Shader getting killed in TravisCI containers.
In those containers, 4g of memory is available but the jvm detects the
full host ram and so defaults to 16g as the default XmX. With no
jvm_options being plumbed to the Shader (they were already to the
underlying IvyUtils and JarTask) - it happily bombed past the container
memory and was OOMkilled.