Export: support export of sources and globs
Add information about sources and/or source globs to the information
returned by pants export.  This could be useful for generating
sparse-checkout files for git.  The glob information is,
unfortunately, somewhat imprecise; it loses information about
excludes.  That's because that data is not always possible to
linearize into a list of (globs and negative globs), which is what
sparse-checkout requires.

Also add the ability to skip library information, so that this
information can be generated more quickly.