Provide global option to look up BUILD files in git history
The ultimate goal here is to support a pants filedeps command that can
generate the necessary data for git sparse checkouts of the just the
data necessary to build a specific target.  To do this, we need to be
able to use the filedeps goal with BUILD files that we haven't checked out
(since we don't want to have to check out all the BUILD files only to
throw most of them away).

The changes to the java_tests target are necessary because otherwise
we'll crash while generating the sparse-checkout data if we don't
happen to have any source files checked out yet.  Of course, we could
make it, too, operate on the in-repo data, but that would be a massive
amount of complexity.  Instead, we simply wait until test time to
complain.