Change ZincAnalysis{Parser} to be a thin wrapper around external zincutils.
I split the zinc parse/split/merge logic out into pantsbuild/zincutils so we
can easily iterate on optimizing it. In particular, if we want to replace part/all
of it with native code, that will be harder to do in pantsbuild/pants, as pants
doesn't yet support building and packaging native code.

This commit turns ZincAnalysis and ZincAnalysisParser into adapters that wrap
that standalone code and adapt it to pants's Analysis/AnalysisParser class hierarchy.

In the future we can consider moving the zinc split/merge code back into the mainline
pantsbuild/pants repo, once we're happy with performance and have an obvious way to
build/deploy it.