Use jvm-compilers as the parent of isolation workunits instead of 'isolation', add workunits for analysis
Before the execution graph's time was under isolation for each of the compiler types. This made it harder to see what was going on because the context managed by isolation was only creating the pool of workers. Moving them to have jvm-compilers as a parent means that as each compile runs, it gets appended to the same list even if it is using a different task. This makes it easier to see things over time in the html reporter output

Testing Done:
Ran locally with changes while watching the html output. CI baking on PR.

Bugs closed: 1464

Reviewed at https://rbcommons.com/s/twitter/r/2134/