propagate keyboard interrupts from worker threads
If a worker thread happens to be the active thread, Ctrl-C will kill it, but not the pants process because the interrupt isn't propagated to the main thread. This changes the WorkerPool s.t. it's work definition will intercept a keyboard interrupt and interrupt main.

Testing Done:
Wrote a regression test, made it pass, CI baking on PR.

Bugs closed: 1473

Reviewed at https://rbcommons.com/s/twitter/r/2143/