Provide global option to look up BUILD files in git history
The ultimate goal here is to support a pants filedeps command that can
generate the necessary data for git sparse checkouts of the just the
data necessary to build a specific target.  To do this, we need to be
able to use the filedeps goal with BUILD files that we haven't checked out
(since we don't want to have to check out all the BUILD files only to
throw most of them away).

The changes to the java_tests and jvm_app targets are necessary
because otherwise we'll crash while generating the sparse-checkout
data if we don't happen to have any source files checked out yet.  Of
course, we could make those, too, operate on the in-repo data, but
that would be a massive amount of complexity.  Instead, we simply wait
task run time to complain.

Testing Done:
CI green: https://travis-ci.org/pantsbuild/pants/builds/60741897

Reviewed at https://rbcommons.com/s/twitter/r/2121/