Remove zincutils tests.
They are a subset of the tests in pantsbuild/zincutils, and it's
pointless and annoying to maintain them in both places.

If/when we move that code back into this repo we will move the tests
back too of course.

Testing Done:
CI passes: https://travis-ci.org/pantsbuild/pants/builds/60866256

Reviewed at https://rbcommons.com/s/twitter/r/2149/