Make a ScalaPlatform subsystem.
This replaces scala/TargetPlatform, and its config dependencies.

It allowed/required some refactoring of ZincUtils, notably:
  - Moving the code that turns jvm tool specs into targets
    (for fingerprinting/invalidation) into JvmToolTaskBase.
    There's nothing zinc-specific about that logic anyway.
  - Moving scalac plugin-related logic into ScalaCompile.

We've been meaning to slim down (and maybe kill entirely) ZincUtils,
so this was a good thing to do anyway.  Now a lot of that code is
more naturally placed alongside code it uses, instead of ZincUtils
reaching back to ScalaCompile to reference things.

Also adds support for subsystems in group tasks, which was overlooked
previously.